Update to use revised version of the font
This is a revised version of the v2 font introduced in 0f0dc41.

The vertical metric for the woff and woff-2 web fonts has now been adjusted in order to center the cap height vertically in a html element (for all browsers) as we’re seeing alignment issues across all browers on Windows.

The previous version of the ‘v2’ font has not been released yet.