Simplify focus styles
Avoids multiple :focus rulesets, and applies the additional outline focus states only when the tab interface is used in IE8 (i.e. when JS is enabled) as that’s the only time its relevant – this means that the no-JS experience is consistent with how we style links elsewhere.

This also defines the focus state outside of the `--selected` modifier because there are weird edge cases where a tab can be focused but not selected (e.g. moving from the mobile view to the desktop view with a link other than the current tab focused)