Promote visual styling of tab to list item
The ‘tab styling’ is currently applied to the link, with the list item doing very little in terms of presentation.

This is problematic when it comes to presenting the focus state, which we want to be tight around the tab text, and thus visually smaller than the tab.

By moving the tab styling to the list item, we can then make the link element smaller, and use an `::after` pseudo element to make the target area for the link fill the entire tab.