Add example of fixed-width header w/ product name
The only examples we currently have for headers with product names are all full width.

We know that in many cases a product name will be used without the header being full width, so we should support this.

At the minute, this example does not render as we’d want it to – the ‘Product Name’ wraps onto a new line – so we’ll need to look at fixing this in the future.