Only polyfill details component if native does not exist
This change simplifies the polyfilling we do for details to only polyfill if the browser does not support the native details element.

I have tested this in many different variations of assistive technologies and browsers and this seems to be a positive change.

iOS VoiceOver goes from: "help with nationality, button" to "help with nationality, collapsed"
"double tap to expand"

One arguably negative is that in JAWs you can no longer use aria-controls functionality to move to the summary details, but given that the content is right under the summary I don't think this would cause a barrier.

All this in mind, this feels like a really good future friendly change that'll allow GOV.UK Frontend to take advantage of native support for this element going forwards.