Explicitly set font weight of warning-text
Explicitly set the font-weight of the warning text component, and move the font styles to the __text element (the __icon element has its own size, weight and colour)

Warning text is otherwise only bold because of the user-agent styling of <strong> elements. When used with GOV.UK Elements, which includes a reset stylesheet, the font-weight is reset to normal, which means the warning text is currently set in a normal weight.