Fix height and alignment issue within header
A change in Chrome 76 means that the header is currently being rendered 64px high (rather than the expected 60px), with mis-alignment between the crown and the GOV.UK logotype.

Adding `display: inline-block` to the __logotype element fixes it.

(We are not sure what exactly changed in Chrome 76, nor why this fixes it.)

Co-authored-by: Trevor Saint <trevor.saint@digital.cabinet-office.gov.uk>