Do not output empty conditionals
If `item.conditional.html` is falsy, perhaps because it’s referencing a variable that is conditionally assigned, the macro currently outputs an empty `div.govuk-checkboxes__conditional` and associates it with the input using aria-controls.

By testing for `item.conditional.html` being truthy (which is the only property in the item.conditional object used within the macro) we can avoid this.