Make count message required
The count message always exists (since 7b534bf). This PR makes the markup for it required,
instead of giving the option to add it with JS since the user needs to be informed
of the limit in the number of characters they can type in even if JS doesn’t run.

The count message DOM element is now defined at the top of the module with the other
required DOM elements. We check whether the textarea is present before using its id to
find the count message - the check prevents a DOM search error for `$textarea.id` if the
script is run without component markup being present present on the page.