Update hex value for secondary text to improve contrast
Fixes in this PR

The secondary text colour value has been updated to improve the
contrast of components that use the ```govuk-colour("dark-grey")```
sass variable. The value has been updated from ```#6f777b``` to ```#626a6e```

This is a small incremental change that improves the contrast ratio from 4.56:1 to 5.51:1.

Fixes #1134