Add an example of creating a custom page width
Set it on the template, header and footer containers.

I’ve used BEM for the custom class name as it feels clearer and in the spirit of
what this is doing (creating something based on the default width class), although
actually nothing is being inherited from `.govuk-width-container`.

There’s a difference in how the header and footer treat `.govuk-width-container`:
footer hardcodes the class whereas header overrides it with `containerClasses`. As
nothing is being inherited from `.govuk-width-container` don't pass it again
unnecessarily, only `.govuk-width-container—wide`.