Fix classes on character count when in error state
The template correctly adds whitespace between classnames when either the `—error` modifier is added or when custom classes are added, but not when both are added.

This means for example that if you were passing the custom class `app-character-count--custom-modifier` to the character count component whilst it was in an error state, the resulting class attribute would be `govuk-js-character-count govuk-textarea--errorapp-character-count--custom-modifier` – neither the error styling nor the custom class would be correctly applied.

Fixes #1630