Add display override for hiding content when printing
When printing users often need to print the content without the surrounding site furniture (eg menus, breadcrumbs, back links and footers).

Introduce a helper so that a class can be used to hide any given element from the print view.

This is explicit and preferable to listing all the selectors for things that shouldn't be printed. These lists grow unwieldy over time.