Test Sass compilation against supported compilers
Compile the GOV.UK Frontend Sass using the minimum version of each of the Sass compilers that we support:

- LibSass v3.3.0
- Dart Sass v1.0.0
- Ruby Sass v3.4.0 (deprecated; not recommended)

Fail the build if it fails to compile using any one of them.

We're doing this so that we have the confidence to ship releases whilst knowing that they should work with any of the Sass compilers that we support.

This also removes a before_deploy step (because we no longer have a deploy step) and the sudo key as according to Travis' config linter it has no effect any more.

Closes #1684