Use aria-describedby relationship for count message
Since aria-describedby is required for the hint message to announced by assistive technologies when interacting with the textarea we can use it to find the count message.

This makes the JavaScript slightly less reliant on the HTML for the component, for example if someone wants to use a different ID on their count message.