Remove unused dependencies
gulp-data was used to generate README files automatically, introduced in 894ba63.

gulp-to-markdown was then used to create the README files as markdown, introduced in 5ac9d18.

gulp-nunjucks was added in dd0bd6b in order to compile the Nunjucks templates to HTML for the READMEs.

The functionality to create READMEs automatically was removed in b7dcac2, but the dependencies were not.