Split component styles from dependencies
Add a new file called _index.scss to each component, which contains the component’s styles but does not import the settings, helpers and tools layers.

Update the existing _[component].scss files to import the settings, helpers and tools layers and import _index.scss.

This will preserve the existing behaviour, and continue to provide a single entry point for individual components that can be used in contexts such as React.

We already have a test in src/govuk/components/all.test.js which ensures that the _[component].scss file renders without errors.

We're doing this so that we can reduce the number of imports within the codebase – by moving everything except the imports into a separate file (`_index.scss`) and including that in the original _[component].scss file, we can then import just the _index.scss file when importing all components, shortcutting all of the base layer imports.