Check that checkbox aria-controls is a conditional
Check that the checkbox aria-controls target has the `govuk-checkboxes__conditional` class before trying to toggle the `--hidden` modifier.

This mirrors a similar check that we do for conditional reveals associated with radios, and makes it more robust if e.g. aria-controls is being used for things that are not conditional reveals, now that we're not scoping the aria-controls lookup to the $module.