Use better 'autocomplete' example for date input
You'd never use "postal-code" or "username" as an autocomplete attribute on an input in the date input component, but you might use "bday-day" if the date input was asking for the user's birthday.

Provide "bday-day" as a more realistic example to help users understand how the attribute can be used effectively with this component.