Add govuk-frontend CSS ordering rules
This ports the rules from config/.sass-lint.yml [1] for ordering CSS
properties. This ordering rule seems to be distinct to govuk-frontend
and not adopted in other parts of GDS (as far as I'm aware) so I've set
this rule as an extension specifically for govuk-frontend. There is
probably a conversation to happen at some point as to whether this
should be adopted wider or whether govuk-frontend wish to continue with
it.

[1]: https://github.com/alphagov/govuk-frontend/blob/2f73bd6d639db10c61d3b0da6f79f08150a4e530/config/.sass-lint.yml#L250-L405