Fix checkbox if border-box sizing applied globally
If a user applies a global box-sizing reset [1], the checkmarks shrink towards the top left corner of the box.

The ::before pseudo-element already uses `box-sizing: border-box`, so update the `::after` pseudo element to use border-box sizing model in order to be consistent. As the box-sizing model is now set explicitly, it should display correctly whether the reset is applied or not.

Increase the width and height of the pseudo elements to include the border widths to accomodate the width of the border (adding an additional 5px to the width and height of normal-sized checkboxes, and 3px to smaller ones).

Switch out the use of `$govuk-border-width` for a hardcoded 5px width for the check marks, as we don't currently factor this in when positioning the check mark, which means if you change the border width setting the checkbox will no longer be centered (I also just don't think it makes sense to tie the width of the check marks to the border width?)

[1]: https://css-tricks.com/inheriting-box-sizing-probably-slightly-better-best-practice/