Turn on selector-pseudo-element-colon-notation stylelint rule
This enforces that psuedo elements are defined with a single colon. In the
config/.sass-lint.yml file the rule for pseudo-elements was disabled which
meant this project has a mix of both single colon and double colon
pseudo elements.

The rule to use single colon rather than double was requested by GOV.UK
frontend developers as a means to not purposely break IE8 despite a lack
of a official support, see
https://github.com/alphagov/stylelint-config-gds/pull/1 for more
information.