Deprecate $govuk-border-width-form-element-error
The border width of form elements in their error state (`$govuk-border-width-form-element-error`) is now the same as form elements in their normal state (`$govuk-border-width-form-element`) – both are 2px.

This isn't something we think users need to be able to configure, and we don't really want the complexity of trying to support different border widths for the error state (for example, the box shadow override that we removed in 6bdca06 would need to be conditionally added if the user changed $govuk-border-width-form-element in order to maintain a consistent focus state).

We'll remove this setting entirely in v4.0, the next breaking release.