Swap modern and legacy palettes in colour settings
I refer to the colour palette file quite a lot in order to grab colour hex values (for example, to check colour contrast combinations or use them in presentations).

Because the legacy colour palette appears first, it's easy to accidentally grab the colour from the legacy palette rather than the current colour palette.

Swap the order so that the current colour palette appears first in the file.