Update stylelint disable comments for Stylelint 13.7.0
In stylelint 13.7.0 they now treat multiline SCSS comments as a single
comment [1] which means the disabling techniques we previously used no
longer work. By adopting this new syntax the reasons are also understood
to be associated with the rule and can allow switching on the
reportDescriptionlessDisables option [2] which can require the disabling
of rules to always be documented.

For example

```
// comment describing disabling
// stylelint-disable indentation
```

is treated as a comment of:
"comment describing disabling stylelint-disable indentation"

Instead this can be resolved with the following technique:

```
// stylelint-disable indentation -- comment describing
// disabling

```

This therefore updates all of the stylelint disabling of rules to match
this pattern.

[1]: https://github.com/stylelint/stylelint/issues/4886