Define visibility once and sync state together
'Flipping' the different state attributes individually allows for them to get out of sync with each other.

Instead, define the visible 'state' once, based on the return value from classList.toggle (true if the `--open` modifier is now present, false if it's removed) and set the other attributes based on this.