Expose aria attributes when initialising
Expose the hidden / collapsed state of the navigation on page load, rather than only after the menu button to be interacted with for the first time.

Find the menu button and menu when the module is initialised and store them in the instance so we can easily sync them without having to pass references to the button and menu between functions.

Add tests to check the aria state of the menu and menu button when the page is first loaded.