Add print styles for the panel component
Because the panel component is currently white text on a green background, it prints without the background with with light grey text (at least in Chrome).

Add print styles to remove the background, set the text colour to black and add a border (like we do when colours are overridden).