Mark header navigation.text and navigation.html as required
It is possible to pass navigation items without a `href` - you might want to do this if you're showing the current page.

However, passing a navigation item without one of `text` or `html` results in no text or link rendering at all.

Following the pattern we use for other components, we should mark `navigation.text` and `navigation.html` as required and use the description to explain that one overrides the other.

The description already explains the relationship between the two. Therefore, we should mark the navigation.text and navigation.html as required.