Mark navigation.text and navigation.href as required
Both `text` and `href` attributes need to be present to create a link.

A navigation item with no href or text does not render - you end up with an empty <ul> element. We also instruct users that a navigation.title is required, so you also end up with a heading with an empty list underneath.

Therefore, mark navigation.items.text and navigation.items.href as required