Change after-build-package tests to look at all objects
The way the previous tests were structured seemed to test for the presence of at least one single object within the component options with the required attributes, instead of all options having the correct attributes. The same was happening for the fixtures file.

This meant that one missing `required` attribute in a component yaml file would not cause the test to fail.

This commit tests for the presence of the required attributes by looping through each option/fixture. It also fixes one typo picked up by these new tests.