Add reduced example of conditional reveals in context
In order to properly test the behaviour of conditional reveals you need to place them in a context where they’re not the only thing on the page. For example, if the only thing on the page is a set of radio buttons then the legend is not read out as expected because there are no other focusable controls to shift focus ‘from’.

Add examples of radios and checkboxes in the context of a basic page, with an additional focusable link before each group of controls.