Change visited state for error summary and notification banner
We recommend not having a different visited link state where it is not helpful to distinguish between visited and unvisited states, for example when linking to pages with frequently-changing content such as the dashboard for an admin interface.

It doesn't seem useful to highlight visited links in the notification banner, and the previous behaviour for the error summary also matches this.

Therefore this commit updates the error and success link style mixins so that the visited colour is the same as the original link colour.