Set notification banner text colour
The notification banner content currently uses the user agent default text colour.

This means that for users who have not changed their user agent default text colour, the text will be true black (`#000000`) rather than the ‘GOV.UK’ black we specify in the colour palette (`#0b0c0c`).

It also means that if a user has changed their default text colour, but has not chosen to override colours, depending on the user’s prefered colour scheme the text may have low contrast or be invisible. For example, my Firefox instance is configured to use white text on a blue background. The notification banner is therefore white-on-white, and thus invisible. As I understand it, this is as a ‘failure of Success Criterion 1.4.3, 1.4.6 and 1.4.8 due to specifying foreground colors without specifying background colors or vice versa’ [1].

Set the text colour of the `govuk-notification-banner__content` element using the `govuk-text-colour` mixin to fix this.

[1]: https://www.w3.org/TR/WCAG20-TECHS/F24.html