Expand git tag error in build-release script
This tripped me up when I was doing the release - I had forgotten to update the package.json file, but thought the error was about the new release tag already existing.

Hopefully adding the tag itself to the error message and suggesting the package.json check will help if anyone else gets this error in the future.