Ignore changes in package.json and CHANGELOG.md
Our release process involves updating `CHANGELOG.md` [1] and `package/package.json` [2] and then running the build-release script WITHOUT committing those changes, so we expect both of those files to be modified at the point the build-release script runs.

Ignore changes in both of those files by excluding them in the pathspec [3] passed to git status.

Store the output of running git status so that we can include it in the output when there are unstaged changes.

[1]: https://github.com/alphagov/govuk-frontend/blob/9fc38d58e356c9188a16071ae0e3580a94bbd608/docs/releasing/publishing.md#:~:text=Update%20CHANGELOG.md%20%22Unreleased%22%20heading%20with%20the%20new%20version%20number.
[2]: https://github.com/alphagov/govuk-frontend/blob/9fc38d58e356c9188a16071ae0e3580a94bbd608/docs/releasing/publishing.md#:~:text=Update%20package%2Fpackage.json%20version
[3]: https://git-scm.com/docs/gitglossary#Documentation/gitglossary.txt-exclude