Add GitHub Action to diff changes to dist
When a pull request makes changes to files in the dist directory, automatically post a prettified git diff of the changes made to the dist folder.

This broadly follows the same steps that we've been doing manually, as outlined in https://scripter.co/git-diff-minified-js-and-css/:

1. Install js-beautify
2. Define two new diff drivers [1] using `git config` for minified JS and CSS
3. Configure Git to use the diff driver for minified CSS and JS in .gitattributes - if the driver is not defined (for example, because it's not in a contributor's local Git config) it seems to be ignored.
4. Run git diff against the base branch for the PR for the dist folder, whilst allowing for renames using `-M` [2] with a similarity index of 10% which has worked well previously.

The diff is then added to the PR as a comment using a GitHub Script action.

[1]: https://git-scm.com/docs/gitattributes#_defining_an_external_diff_driver
[2]: https://git-scm.com/docs/git-diff#Documentation/git-diff.txt--Mltngt