Use tabular numbers for character count message
By default different numbers take up different amounts of horizontal space, for example ‘1’ is narrower than ‘0’, just like an ‘i’ is narrower than an ‘m’.

This means that in the character count message the amount of space take up by the count varies as the user types. This causes the words after the count (‘characters remaining’) to jump around in a distracting way.

Tabular numbers take up a fixed amount of space, which means the words after the count do not shift position while the count maintains the same number of digits.