Remove redundant align-button modifiers in example
The original intent was to use a modifier on links to align them with buttons, but we ended up finding another approach that used flexbox and didn’t require the modifier. However, the classes were never removed from the example page in the review app.

Remove the redundant classes as they are not doing anything and will only cause confusion.