Add cookie banner styling
The component does not set bottom spacing. The bottom spacing should be created by the
items inside the component.

 We're going to use the `hidden` attribute to hide sections as this will hide the content even if CSS fails to load.

`display: none;` will act as a fallback for legacy browsers (IE 8-10) that don't support the `hidden` attribute.
Unfortunately it looks like Lynx doesn't support `hidden` but that's a vendor implementation issue. It also appears that
Lynx hasn't been in active development for some time. Using the `hidden` attribute seems like the semantically correct
thing to do here and it makes the component usable when CSS fails to load.

(The alternative would be to insert text into the DOM at relevant points which could also be a suitable solution in some
contexts. We decided against it as the Design System components need to be really flexible so we're trying to avoid
patterns where HTML markup needs to be maintained inside JavaScript which will also simplify localisation etc. where
relevant.)

Visually separate the cookie banner from content underneath with a transparent bottom border
when user changes colours in their browser.