Cursor to pointer as per other links
New CSS to change the cursor to pointer when hovered over, as per any other link, including other links in the header. An edge case perhaps because this will presumably mostly be for mobile, but for desktops with a narrow window, or heavily zoomed in, a pointer cursor will be consistent with other links.