Enable cookie banner action to be a link button
Enable the action to be a link styled as a button, with role="button" and draggable="false", if both
href and type="button" are provided. This matches the [Start buttons](https://design-system.service.gov.uk/components/button/#start-buttons).

( A more terse solution to achieve this in the macro would be to skip the first condition with

```
{% if action.href && action.type != "button"% }
```

and pass a `href` to the button macro in `else`. This would work since Nunjucks doesn't render blank
attributes for options whose values aren't specified in the macro. However, other templating languages
might not follow a similar pattern - and we should try and keep our macro aligned with what ports of
GOV.UK Frontend do in their macros.)