Add an example to the review app of essential cookies
Add an example of a typical essential cookies only banner that works both
on client and server side. When the user hides the cookie banner, we
progressively enhance the cookie banner to hide with JavaScript. If JavaScript
is not enabled, the form is submitted and a page navigation occurs.

To enable progressive enhancement, we include a small amount of client-side
JavaScript to handle the ‘Hide’ button in the example. The cookie banner component
itself does not ship with JavaScript. Typically the service team needs to write the
code that hides the banner and integrates the collection of cookies both in JavaScript
and on the server.