Tidy up hover/focus state for logo w/ product name
When the header includes a product name, the homepage link’s underline (on hover) and the focus state take up an additional 10px on the right hand side, because of the 10px right padding on `govuk-header__product-name`.

This padding is there to ensure there is a gap left between the product name and the menu button [1]. Remove it, and add a 10px margin to `govuk-header__link--homepage` instead which achieves the same result but without adding an additional 10px to the hover and focus states.

[1]: https://github.com/alphagov/govuk-frontend/commit/df413cf6363b537fe37b1b0db5bc40b97424422f