Simplify back link styles
We used to use a bottom border to create the underline so that the ‘triangle’ and the link were both underlined.

This was changed in ad141f4 to make it consistent with breadcrumbs, which means the special link treatment here isn’t required any more.

Remove it, in preparation for the new link styles which it will otherwise override.