Update tabs to use core link mixin
Tabs are progressively enhanced from a set of links, which is what users see if JavaScript is not enabled.

Apply the common link styles mixin, which includes the new link style treatments, to the tabs. The common link styles mixin also includes the focus state, so we no longer need to apply that separately.

When we progressively enhance the tabs, the only thing we then need to change about the links is the link colour, which is already handled by the inclusion of the `govuk-link-style-text` mixin.

Using the text link style mixin for this component prepares us for the new link styles which will now be inherited from the mixins.